DC_request_24_June_newIP/requirements.txt
pandas==0.25.3
scikit-learn==0.22
paho-mqtt
cryptography
asymcrypt
